Allow aliases in any order
Before, if one alias was a substring of another, they had to be ordered
with the longest first, e.g.

    aliases: { 4: ['fore', 'for'] }

Otherwise, the input string could contain the longer alias but the
parser would match it to the shorter alias, leaving extra unmatched
characters on the end. This commit allows the user to specify aliases in
any order.